Move findSimilarRecipes function to recipeUtils module for better organization
